restful_api/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.4
django-filter==1.0.4
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
httpie==0.9.9
idna==2.5
Markdown==2.6.8
olefile==0.44
Pillow==4.2.1
Pygments==2.2.0
PyJWT==1.5.2
pytz==2017.2
requests==2.18.3
urllib3==1.22
whitenoise==3.3.0
